Create an abstract representation of the Twitter friendship graph

Signifies that aPPR should query the Twitter friendship graph via neocache.

Usage

neocache_graph(cache_name = "aPPR", retryonratelimit = TRUE)

Arguments

cache_name

The name of the neocache to use to save data into.

retryonratelimit

If TRUE, and a rate limit is exhausted, will wait until it refreshes. Most Twitter rate limits refresh every 15 minutes. If FALSE, and the rate limit is exceeded, the function will terminate early with a warning; you'll still get back all results received up to that point. The default value, NULL, consults the option rtweet.retryonratelimit so that you can globally set it to TRUE, if desired.

If you expect a query to take hours or days to perform, you should not rely soley on retryonratelimit because it does not handle other common failure modes like temporarily losing your internet connection.

Examples

if (FALSE) {

test_ids <- c("780429268866052096", "1191642560")

graph <- neocache_graph()

check(graph, test_ids)
node_degrees(graph, test_ids)
neighborhood(graph, test_ids[1])

}