R6 class to manage personalized pagerank calculations

R6 class to manage personalized pagerank calculations

R6 class to manage personalized pagerank calculations

Details

Blargh blarh blargh

Public fields

stats

A [tibble::tibble()] with one row for each visited node and the following columns: - `name`: Name of a node (character). - `r`: Current estimate of residual for a node. - `p`: Current estimate of the pagerank for a node. - `in_degree`: Number of incoming edges to a node. - `out_degree`: Number of outcoming edges from a node.

Methods

Public methods

Tracker$new()

Tracker$print()

Tracker$remaining()

Tracker$in_tracker()

Tracker$add_node()

Tracker$update_p()

Tracker$update_r_neighbor()

Tracker$update_r_self()

Tracker$clone()

Method new()
Create a new Tracker object.
Usage

Tracker$new()

Returns
A new `Tracker` object.

Method print()
Print the tibble containing the current state of the pagerank calculation.
Usage

Tracker$print()

Method remaining()
Determine nodes that need to be visited
Usage

Tracker$remaining(epsilon)

Arguments

epsilon

The error tolerance / convergence parameter in Algorithm 3.

Returns
A character vector of node names with current residuals greater than `epsilon`.

Method in_tracker()
Check if there is already a row for a particular node
Usage

Tracker$in_tracker(node)

Arguments

node

Character name of a node in the graph.

Returns
`TRUE` if there is a row for `node`, `FALSE` if there is not a row for `node`.

Method add_node()
Create an entry for `node` in the tracker. Assumes that `node` is not in the tracker yet, and does not check if this is the case.
Usage

Tracker$add_node(graph, node, preference = 0)

Arguments

graph

The graph object.

node

The name of the node in the graph as a length 1 character vector.

preference

TODO: recall what on earth this is.

Method update_p()
Update the estimate of the personalized pagerank for a given node
Usage

Tracker$update_p(node, alpha_prime)

Arguments

node

Character name of a node in the graph.

alpha_prime

Transformed teleportation constant from Algorithm 3.

Method update_r_neighbor()
Update the residual of a node in the neighborhood of current node
Usage

Tracker$update_r_neighbor(graph, u, v, alpha_prime)

Arguments

graph

The graph object.

u

Character name of the node we are currently visiting.

v

Character name of a neighborhor of `u`.

alpha_prime

Transformed teleportation constant from Algorithm 3.

Method update_r_self()
Update the residual of current node
Usage

Tracker$update_r_self(node, alpha_prime)

Arguments

node

Character name of the node we are currently visiting.

alpha_prime

Transformed teleportation constant from Algorithm 3.

Method clone()
The objects of this class are cloneable with this method.
Usage

Tracker$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.