Approximate personalized pageranks

Approximate personalized pageranks

appr(graph, seeds, alpha = 0.15, epsilon = 1e-06, tau = NULL, ...)

# S3 method for igraph
appr(graph, seeds, alpha = 0.15, epsilon = 1e-06, tau = NULL, ...)

# S3 method for rtweet_graph
appr(graph, seeds, alpha = 0.15, epsilon = 1e-06, tau = NULL, ...)

Arguments

graph
An [abstract_graph()] object, such as those created by [rtweet_graph()] and [twittercache_graph()], and [igraph::igraph()] object. This argument is required.

seeds
A character vector of seeds for the personalized pagerank. The personalized pagerank will return to each of these seeds with probability `alpha` at each node transition. At the moment, all seeds are given equal weighting. This argument is required.

alpha
Teleporting factor. The teleportation factor is the probability of returning to a seed node at each node transition. Defaults to `0.15`. This is the inverse of the "dampening factor" in the original PageRank paper, so `alpha = 0.15` corresponds to a dampening factor of `0.85`.

epsilon
Desired accuracy of approximation. Small `epsilon` can result in long runtimes. Defaults to `1e-6`.

tau
Regularization term. Additionally inflates the in degree of each observation by this term by performing the degree adjustment described in Algorithm 3 and Algorithm 4, which are described in `vignette("Mathematical details")`. Defaults to `NULL`, in which case `tau` is set to the average in degree of the observed nodes. In general, setting it's reasonable to set `tau` to the average degree of the graph.

...
Ignored. Passing arguments to `...` results in a warning.

Value

A [tibble::tibble()] with the following columns:

- `name`: Name of a node (character). - `p`: Estimated personalized pagerank of a node. - `r`: Estimated error of pagerank estimate for a node. - `in_degree`: Number of incoming edges to a node. - `out_degree`: Number of outcoming edges from a node. - `degree_adjusted`: The personalized pagerank divided by the node in-degree. - `regularized`: The personalized pagerank divide by the node in-degree plus `tau`.

When `graph` is an

- `rtweet_graph` get `user_id` not screen names

Details

Note that, due to the inspection paradox, the average observed degree of the network will be higher than the average degree of the entire network. This means that the default value for `tau` many be a bit high. In practice, we do not observe that the small changes in the value of `tau` substantively changes results, you really just need to regularize enough so that low in degree nodes don't look overly important after degree correction.

Examples

library(aPPR)
library(igraph)

#> 
#> Attaching package: 'igraph'

#> The following object is masked from 'package:testthat':
#> 
#>     compare

#> The following objects are masked from 'package:stats':
#> 
#>     decompose, spectrum

#> The following object is masked from 'package:base':
#> 
#>     union

set.seed(27)

graph <- rtweet_graph()
# NOT RUN {
appr(graph, "alexpghayes")
# }

graph2 <- sample_pa(100)

appr(graph2, seeds = "5")

#> # A tibble: 3 x 7
#>   name            r     p in_degree out_degree degree_adjusted regularized
#>   <chr>       <dbl> <dbl>     <dbl>      <dbl>           <dbl>       <dbl>
#> 1 5     0.000000833 0.150         2          1         0.0750      0.0105 
#> 2 4     0.000000579 0.127         4          1         0.0319      0.00781
#> 3 1     0           0.108        31          0         0.00350     0.00250

library(igraph)

set.seed(27)

graph <- sample_pa(100)

appr(graph, seeds = "5")

#> # A tibble: 3 x 7
#>   name            r     p in_degree out_degree degree_adjusted regularized
#>   <chr>       <dbl> <dbl>     <dbl>      <dbl>           <dbl>       <dbl>
#> 1 5     0.000000833 0.150         2          1         0.0750      0.0105 
#> 2 4     0.000000579 0.127         4          1         0.0319      0.00781
#> 3 1     0           0.108        31          0         0.00350     0.00250